Revert "Merge pull request #6 from Qwestive/auth-manager-button"
This reverts commit 12cc292f462319a78fecebc48ed50b483a4b9d61, reversing
changes made to d7e868e9790ccb594878bff28a0035807a625d3a.
