--- src/common/components/SolanaWallet/WalletButton.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { WalletMultiButton as ReactUIWalletMultiButton } from '@solana/wallet-adapter-react-ui';
-
-/*
-TODO:
-- make a prettier button
-*/
-export default function WalletButton(): JSX.Element {
-  return <ReactUIWalletMultiButton />;
-}
